+++ AppboyProject/android/app/src/main/java/com/appboyproject/MainApplication.java
@@ -6,6 +6,7 @@ import android.util.Log;
 import com.appboy.AppboyLifecycleCallbackListener;
 import com.appboy.reactbridge.AppboyReactPackage;
 import com.appboy.support.AppboyLogger;
+import com.facebook.react.PackageList;
 import com.facebook.react.ReactApplication;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
@@ -25,10 +26,9 @@ public class MainApplication extends Application implements ReactApplication {
 
     @Override
     protected List<ReactPackage> getPackages() {
-      return Arrays.<ReactPackage>asList(
-          new MainReactPackage(),
-          new AppboyReactPackage()
-      );
+      @SuppressWarnings("UnnecessaryLocalVariable")
+      List<ReactPackage> packages = new PackageList(this).getPackages();
+      return packages;
     }
   };